Altered Travis build matrix
It looks like mpicc is intent on using the main version of gcc on the
build box (gcc 4.8) for now, so until we work that out, I'm trimming the
build matrix down.

Also, the configure script appears to not try to detect HDF5, so I'm
adding an extra configure flags variable to enable the HDF5 interface
use.